Moved modalData to Data folder, Backdrop files to Modal in Atoms, improved imports(paths)
